Fix `toJsonString()` API
