+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/utils/JsonUtils.java
@@ -17,17 +17,29 @@
  */
 package io.ballerina.runtime.api.utils;
 
+import io.ballerina.runtime.api.PredefinedTypes;
+import io.ballerina.runtime.api.TypeTags;
+import io.ballerina.runtime.api.creators.TypeCreator;
+import io.ballerina.runtime.api.creators.ValueCreator;
+import io.ballerina.runtime.api.types.ArrayType;
 import io.ballerina.runtime.api.types.JsonType;
 import io.ballerina.runtime.api.types.MapType;
 import io.ballerina.runtime.api.types.StructureType;
+import io.ballerina.runtime.api.types.TableType;
 import io.ballerina.runtime.api.types.Type;
 import io.ballerina.runtime.api.values.BArray;
 import io.ballerina.runtime.api.values.BError;
+import io.ballerina.runtime.api.values.BIterator;
 import io.ballerina.runtime.api.values.BMap;
+import io.ballerina.runtime.api.values.BRefValue;
 import io.ballerina.runtime.api.values.BString;
 import io.ballerina.runtime.api.values.BTable;
 import io.ballerina.runtime.internal.JsonGenerator;
 import io.ballerina.runtime.internal.JsonParser;
+import io.ballerina.runtime.internal.TypeChecker;
+import io.ballerina.runtime.internal.commons.TypeValuePair;
+import io.ballerina.runtime.internal.util.exceptions.BLangExceptionHelper;
+import io.ballerina.runtime.internal.util.exceptions.RuntimeErrors;
 import io.ballerina.runtime.internal.values.ErrorValue;
 
 import java.io.IOException;
@@ -36,6 +48,13 @@ import java.io.OutputStream;
 import java.io.Reader;
 import java.io.Writer;
 import java.nio.charset.Charset;
+import java.util.List;
+import java.util.Map;
+
+import static io.ballerina.runtime.api.creators.ErrorCreator.createError;
+import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.VALUE_LANG_LIB_CONVERSION_ERROR;
+import static io.ballerina.runtime.internal.util.exceptions.BallerinaErrorReasons.VALUE_LANG_LIB_CYCLIC_VALUE_REFERENCE_ERROR;
+import static io.ballerina.runtime.internal.util.exceptions.RuntimeErrors.INCOMPATIBLE_CONVERT_OPERATION;
 
 /**
  * Class @{@link JsonParser} provides APIs to handle json values.
@@ -256,6 +275,114 @@ public class JsonUtils {
         } catch (IOException e) {
             throw new ErrorValue(StringUtils.fromString(e.getMessage()), e);
         }
+    }
+
+    public static Object convertToJson(Object value, List<TypeValuePair> unresolvedValues) {
+        Type jsonType = PredefinedTypes.TYPE_JSON;
+
+        if (value == null) {
+            return null;
+        }
+
+        Type sourceType = TypeChecker.getType(value);
+
+        if (sourceType.getTag() <= TypeTags.BOOLEAN_TAG && TypeChecker.checkIsType(value, jsonType)) {
+            return value;
+        }
+
+        TypeValuePair typeValuePair = new TypeValuePair(value, jsonType);
+
+        if (unresolvedValues.contains(typeValuePair)) {
+            throw createCyclicValueReferenceError(value);
+        }
+
+        unresolvedValues.add(typeValuePair);
+
+        Object newValue;
+        switch (sourceType.getTag()) {
+            case TypeTags.XML_TAG:
+            case TypeTags.XML_ELEMENT_TAG:
+            case TypeTags.XML_COMMENT_TAG:
+            case TypeTags.XML_PI_TAG:
+            case TypeTags.XML_TEXT_TAG:
+                newValue = StringUtils.getStringValue(value, null);
+                break;
+            case TypeTags.TUPLE_TAG:
+            case TypeTags.ARRAY_TAG:
+                newValue = convertArrayToJson((BArray) value, unresolvedValues);
+                break;
+            case TypeTags.TABLE_TAG:
+                BTable bTable = (BTable) value;
+                Type constrainedType = ((TableType) bTable.getType()).getConstrainedType();
+                if (constrainedType.getTag() == TypeTags.MAP_TAG) {
+                    newValue = convertMapConstrainedTableToJson((BTable) value, unresolvedValues);
+                } else {
+                    try {
+                        newValue = io.ballerina.runtime.internal.JsonUtils.toJSON(bTable);
+                    } catch (Exception e) {
+                        throw createConversionError(value, jsonType, e.getMessage());
+                    }
+                }
+                break;
+            case TypeTags.RECORD_TYPE_TAG:
+            case TypeTags.MAP_TAG:
+                newValue = convertMapToJson((BMap<?, ?>) value, unresolvedValues);
+                break;
+            case TypeTags.ERROR_TAG:
+            default:
+                throw createConversionError(value, jsonType);
+        }
+
+        unresolvedValues.remove(typeValuePair);
+        return newValue;
+    }
+
+    private static Object convertMapConstrainedTableToJson(BTable value, List<TypeValuePair> unresolvedValues) {
+        BArray membersArray = ValueCreator.createArrayValue(PredefinedTypes.TYPE_JSON_ARRAY);
+        BIterator itr = value.getIterator();
+        while (itr.hasNext()) {
+            BArray tupleValue = (BArray) itr.next();
+            BMap mapValue = ((BMap) tupleValue.get(0));
+            Object member = convertMapToJson(mapValue, unresolvedValues);
+            membersArray.append(member);
+        }
+        return membersArray;
+    }
+
+    private static Object convertMapToJson(BMap<?, ?> map, List<TypeValuePair> unresolvedValues) {
+        BMap<BString, Object> newMap =
+                ValueCreator.createMapValue(TypeCreator.createMapType(PredefinedTypes.TYPE_JSON));
+        for (Map.Entry entry : map.entrySet()) {
+            Object newValue = convertToJson(entry.getValue(), unresolvedValues);
+            newMap.put(StringUtils.fromString(entry.getKey().toString()), newValue);
+        }
+        return newMap;
+    }
+
+    private static Object convertArrayToJson(BArray array, List<TypeValuePair> unresolvedValues) {
+        BArray newArray = ValueCreator.createArrayValue((ArrayType) PredefinedTypes.TYPE_JSON_ARRAY);
+        for (int i = 0; i < array.size(); i++) {
+            Object newValue = convertToJson(array.get(i), unresolvedValues);
+            newArray.add(i, newValue);
+        }
+        return newArray;
+    }
+
+    private static BError createConversionError(Object inputValue, Type targetType) {
+        return createError(VALUE_LANG_LIB_CONVERSION_ERROR,
+                BLangExceptionHelper.getErrorDetails(INCOMPATIBLE_CONVERT_OPERATION,
+                        TypeChecker.getType(inputValue), targetType));
+    }
+
+    private static BError createConversionError(Object inputValue, Type targetType, String detailMessage) {
+        return createError(VALUE_LANG_LIB_CONVERSION_ERROR, BLangExceptionHelper.getErrorMessage(
+                        INCOMPATIBLE_CONVERT_OPERATION, TypeChecker.getType(inputValue), targetType)
+                .concat(StringUtils.fromString(": ".concat(detailMessage))));
+    }
+
+    private static BError createCyclicValueReferenceError(Object value) {
+        return createError(VALUE_LANG_LIB_CYCLIC_VALUE_REFERENCE_ERROR, BLangExceptionHelper.getErrorDetails(
+                RuntimeErrors.CYCLIC_VALUE_REFERENCE, ((BRefValue) value).getType()));
     }
 
     /**
