+++ bvm/ballerina-runtime/src/main/java/io/ballerina/runtime/api/utils/StringUtils.java
@@ -349,31 +349,27 @@ public class StringUtils {
      * @return Json String value of the value
      */
     public static String getJsonString(Object value) {
-        if (value == null) {
-            return "null";
-        }
+        Object jsonValue = JsonUtils.convertToJson(value, new ArrayList<>());
 
-        Type type = TypeChecker.getType(value);
-
-        if (type.getTag() == TypeTags.STRING_TAG) {
-            return stringToJson((BString) value);
+        Type type = TypeChecker.getType(jsonValue);
+        switch (type.getTag()) {
+            case TypeTags.NULL_TAG:
+                return "null";
+            case TypeTags.STRING_TAG:
+                return stringToJson((BString) jsonValue);
+            case TypeTags.MAP_TAG:
+                MapValueImpl mapValue = (MapValueImpl) jsonValue;
+                return mapValue.getJSONString();
+            case TypeTags.ARRAY_TAG:
+                ArrayValue arrayValue = (ArrayValue) jsonValue;
+                return arrayValue.getJSONString();
+            default:
+                if (type.getTag() < TypeTags.NULL_TAG) {
+                    return String.valueOf(jsonValue);
+                }
+                RefValue refValue = (RefValue) jsonValue;
+                return refValue.stringValue(null);
         }
-        if (type.getTag() < TypeTags.NULL_TAG) {
-            return String.valueOf(value);
-        }
-
-        if (type.getTag() == TypeTags.MAP_TAG) {
-            MapValueImpl mapValue = (MapValueImpl) value;
-            return mapValue.getJSONString();
-        }
-
-        if (type.getTag() == TypeTags.ARRAY_TAG) {
-            ArrayValue arrayValue = (ArrayValue) value;
-            return arrayValue.getJSONString();
-        }
-
-        RefValue refValue = (RefValue) value;
-        return refValue.stringValue(null);
     }
 
     private static String stringToJson(BString value) {
